+++ docs/Whats-New.md
@@ -57,9 +57,17 @@ New:
 - Implement voxel light customization (by ZivDero)
 - Implement an option to disable Tiberium storage (by Rampastring)
 
+
 Vinifera fixes:
 - Fix unit placement in non-TS Client builds of Vinifera (by ZivDero)
 - Fix a bug where `CellSpread` incorrectly handled height distance with buildings (by ZivDero)
+
+
+Vanilla fixes:
+- Fix the player being able to unload an aircraft onto a building, resulting in limboed units (port from ts-patches) (by ZivDero)
+- Fix the player being able to land an aircraft onto a helipad even if it's not listed as `Dock=` of this aircraft (by ZivDero)
+- Fix a bug where a weeder could be ordered to dock with a waste facility that wasn't listed in the weeder's `Dock=` key (by ZivDero)
+- Fix a bug where the player couldn't repair aircraft on allied repair bays under some circumstances (by ZivDero)
 
 :::